+++ ntex-bytes/src/lib.rs
@@ -73,7 +73,7 @@ mod string;
 
 pub use crate::bvec::BytesMut;
 pub use crate::bytes::Bytes;
-pub use crate::pages::{BytesPages, Page};
+pub use crate::pages::{BytePage, BytePages};
 pub use crate::string::ByteString;
 
 #[doc(hidden)]
@@ -101,26 +101,29 @@ pub mod info {
     }
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum PageSize {
-    Size8 = 0,
-    Size16 = 1,
-    Size24 = 2,
-    Size32 = 3,
-    Size48 = 4,
-    Size64 = 5,
-    Unset = 6,
+#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
+pub enum BytePageSize {
+    Size4 = 0,
+    Size8 = 1,
+    #[default]
+    Size16 = 2,
+    Size24 = 3,
+    Size32 = 4,
+    Size48 = 5,
+    Size64 = 6,
+    Unset = 7,
 }
 
-impl PageSize {
+impl BytePageSize {
     const fn capacity(self) -> usize {
         match self {
-            PageSize::Size8 => 8 * 1024,
-            PageSize::Size16 => 16 * 1024,
-            PageSize::Size24 => 24 * 1024,
-            PageSize::Size32 => 32 * 1024,
-            PageSize::Size48 => 48 * 1024,
-            PageSize::Size64 | PageSize::Unset => 64 * 1024,
+            BytePageSize::Size4 => 4 * 1024,
+            BytePageSize::Size8 => 8 * 1024,
+            BytePageSize::Size16 => 16 * 1024,
+            BytePageSize::Size24 => 24 * 1024,
+            BytePageSize::Size32 => 32 * 1024,
+            BytePageSize::Size48 => 48 * 1024,
+            BytePageSize::Size64 | BytePageSize::Unset => 64 * 1024,
         }
     }
 }
@@ -131,12 +134,13 @@ mod tests {
 
     #[test]
     fn page_size() {
-        assert_eq!(PageSize::Size8.capacity(), 8 * 1024);
-        assert_eq!(PageSize::Size16.capacity(), 16 * 1024);
-        assert_eq!(PageSize::Size24.capacity(), 24 * 1024);
-        assert_eq!(PageSize::Size32.capacity(), 32 * 1024);
-        assert_eq!(PageSize::Size48.capacity(), 48 * 1024);
-        assert_eq!(PageSize::Size64.capacity(), 64 * 1024);
-        assert_eq!(PageSize::Unset.capacity(), 64 * 1024);
+        assert_eq!(BytePageSize::Size4.capacity(), 4 * 1024);
+        assert_eq!(BytePageSize::Size8.capacity(), 8 * 1024);
+        assert_eq!(BytePageSize::Size16.capacity(), 16 * 1024);
+        assert_eq!(BytePageSize::Size24.capacity(), 24 * 1024);
+        assert_eq!(BytePageSize::Size32.capacity(), 32 * 1024);
+        assert_eq!(BytePageSize::Size48.capacity(), 48 * 1024);
+        assert_eq!(BytePageSize::Size64.capacity(), 64 * 1024);
+        assert_eq!(BytePageSize::Unset.capacity(), 64 * 1024);
     }
 }